feat: extract shared UI into @pi-outpost/ui workspace
Create ui/ workspace following shared/ pattern (raw TSX exports,
no build step). Move 16 files from web/src/:
- 10 components (CopyButton, CodeHighlight, Mermaid, RenderedHtml,
  DiffBlocks, UserMessage, AssistantMessage, ToolCard, CustomMessageCard)
- 3 theme files (ThemeContext, useTheme, theme)
- 4 utility modules (workspacePath, markdownMath, diff, toolOutput)

Update web/App.tsx and web/FileViewer.tsx to import from @pi-outpost/ui.
Add @pi-outpost/ui to web/package.json dependencies.
Add ui to root package.json workspaces.

Closes #16
