+++ package.json
@@ -6,7 +6,8 @@
     "server",
     "web",
     "embed",
-    "cli"
+    "cli",
+    "ui"
   ],
   "scripts": {
     "dev": "concurrently -n server,web -c blue,green \"npm run dev --workspace server\" \"npm run dev --workspace web\"",
