+++ ui/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@pi-outpost/ui",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts",
+    "./components/*": "./src/components/*",
+    "./theme/*": "./src/theme/*",
+    "./util/*": "./src/util/*"
+  }
+}
