+++ ui/src/components/CopyButton.tsx
