+++ ui/src/util/markdownMath.ts
