+++ web/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@pi-outpost/shared": "*",
+    "@pi-outpost/ui": "*",
     "@tailwindcss/vite": "^4.3.2",
     "highlight.js": "^11.11.1",
     "katex": "^0.16.47",
